# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tajweed Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/sync:
    post:
      operationId: syncUser
      tags: [users]
      summary: Sync/provision user after Clerk sign-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSyncInput"
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Courses ─────────────────────────────────────────────────────────────────
...
  /students/{id}:
    get:
      operationId: getStudent
      tags: [users]
      summary: Get student profile (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [users]
      summary: Update student (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  # ── Enrollments ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [teacher, student]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, email, name, role, createdAt]
    UserSyncInput:
      type: object
      properties:
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required: [clerkId, email, name]
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        lastActiveAt:
          type: ["string", "null"]
        totalEnrollments:
          type: integer
        activeEnrollment:
          $ref: "#/components/schemas/Enrollment"
        masteryScore:
          type: integer
        createdAt:
          type: string
      required: [id, clerkId, email, name, totalEnrollments, masteryScore, createdAt]
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # ── Courses ──────────────────────────────────────────────────────────────
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: integer
          description: "1 = Level 1, 2 = Level 2"
        durationDays:
          type: integer
        price:
          type: number
        capacity:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        enrolledCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, level, durationDays, price, capacity, isActive, enrolledCount, createdAt]
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        student:
          $ref: "#/components/schemas/User"
        course:
          $ref: "#/components/schemas/Course"
        createdAt:
          type: string
      required: [id, studentId, courseId, status, createdAt]
...